Remove types file for `MintDurationTiers` component
